saveStructureForDebug: Move into own class
