+++ app/src/main/java/re/neutrino/kanji_assist/AssistSession.java
@@ -5,16 +5,10 @@ import android.app.assist.AssistStructure;
 import android.content.Context;
 import android.graphics.PointF;
 import android.os.Bundle;
-import android.os.Environment;
 import android.service.voice.VoiceInteractionSession;
 import android.util.Log;
 import android.view.MotionEvent;
 import android.view.View;
-
-import java.io.File;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.ObjectOutputStream;
 
 class AssistSession extends VoiceInteractionSession {
 
@@ -48,45 +42,12 @@ class AssistSession extends VoiceInteractionSession {
 
         textExtractor = new TextExtractor(structure);
 
-        saveStructureForDebug(this, structure);
+        new AssistStructureSaver().saveStructureForDebug(structure);
 
         ScreenText selected = textExtractor.getSelectedText();
         if (selected != null) {
             dictionaryPopup.show(selected);
         }
-    }
-
-    private static void saveStructureForDebug(AssistSession assistSession, AnyAssistStructure structure) {
-        if (!BuildConfig.DEBUG)
-            return;
-
-        final FakeAssistStructure fakeStructure =
-                new FakeAssistStructure(structure);
-
-        if (!assistSession.isExternalStorageWritable()) {
-            Log.w("saveStructureForDebug",
-                    "External storage not writable, not saving assist structure");
-            return;
-        }
-
-        // TODO Ask for permission interactively
-
-        File file = new File(Environment.getExternalStoragePublicDirectory(
-                Environment.DIRECTORY_DOCUMENTS), "assist.bin");
-        if (!file.getParentFile().mkdirs()) {
-            Log.e("saveStructureForDebug", "Directory not created, not saving assist structure");
-        }
-
-        try (FileOutputStream output = new FileOutputStream(file)) {
-            new ObjectOutputStream(output).writeObject(fakeStructure);
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-
-    private boolean isExternalStorageWritable() {
-        String state = Environment.getExternalStorageState();
-        return Environment.MEDIA_MOUNTED.equals(state);
     }
 
     private class OnTouchListener implements View.OnTouchListener {
